Move isketch to its own mod
